Add files via upload
This is an analysis of U.S Census Data of 2015 done by tableu software. This is my third project in advanced data analysis track by Udacity and EGFWD Initiative.
You can find summary of the analysis alongside with dashboards and stories links in tableu. Have fun Visualizing and Deriving Insights!